Moving the Error class to utils
